# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateProposalDto": {
        "properties": {
          "notes": {
            "description": "Optional context — reasoning, links, cost estimate, capacity notes.",
            "maxLength": 2000,
            "type": "string"
          },
          "summary": {
            "description": "Short summary of the proposed answer (e.g. \"Lola's Rooftop, Yaba\").",
            "example": "Lola's Rooftop, Yaba",
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "typed": {
            "description": "Optional structured payload (Phase 8+).",
            "type": "object"
          }
        },
        "required": [
          "summary"
        ],
        "type": "object"
      },
      "CreateSegmentDto": {
        "properties": {
          "description": {
            "maxLength": 1000,
            "type": "string"
          },
          "position": {
            "description": "Optional explicit position. Defaults to the end of the list.",
            "minimum": 0,
            "type": "number"
          },
          "segmentType": {
            "description": "Type slug — drives the card icon. Use one of the known set (date_time, venue, budget, agenda, guest_list, roles, theme, travel) for a meaningful icon, or anything else for the generic fallback.",
            "example": "venue",
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          },
          "title": {
            "example": "Catering",
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "title",
          "segmentType"
        ],
        "type": "object"
      },
...
      "LockSegmentDto": {
        "properties": {
          "proposalId": {
            "description": "The proposal that wins. Must belong to this segment and be live with zero live objections.",
            "type": "string"
          }
        },
        "required": [
          "proposalId"
        ],
        "type": "object"
      },
...
      "ProposalProposedByDto": {
        "properties": {
          "email": {
            "example": "amaka@example.com",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "example": "usr_2N9k8x4QyT3pA7c1bWv0eR",
            "type": "string"
          },
          "name": {
            "example": "Amaka Okonkwo",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "ProposalResponseDto": {
        "properties": {
          "anonymous": {
            "example": false,
            "type": "boolean"
          },
          "createdAt": {
            "example": "2026-05-26T10:22:00.000Z",
            "type": "string"
          },
          "id": {
            "example": "a1b2c3d4-e5f6-7890-abcd-ef0123456789",
            "format": "uuid",
            "type": "string"
          },
          "objections": {
            "items": {
              "$ref": "#/components/schemas/ProposalObjectionDto"
            },
            "type": "array"
          },
          "payload": {
            "additionalProperties": true,
            "description": "Shape depends on the parent segment.segmentType — validated server-side per type. E.g. a venue proposal has { name, address, capacity }; a date_time proposal has { date, startsAt }.",
            "example": {
              "address": "12 Marina Way, Lagos",
              "capacity": 150,
              "name": "Garden Court Hotel"
            },
            "type": "object"
          },
          "proposedBy": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProposalProposedByDto"
              }
            ],
            "description": "Null when the proposal is anonymous (event has the 'anonymous_proposals' add-on enabled).",
            "nullable": true
          },
          "proposedById": {
            "example": "usr_2N9k8x4QyT3pA7c1bWv0eR",
            "type": "string"
          },
          "segmentId": {
            "example": "e3b0c442-98fc-4e1b-9c2b-9d3a7c2f1a4d",
            "format": "uuid",
            "type": "string"
          },
          "state": {
            "enum": [
              "live",
              "withdrawn",
              "eliminated",
              "winner"
            ],
            "example": "live",
            "type": "string"
          },
          "updatedAt": {
            "example": "2026-05-26T10:22:00.000Z",
            "type": "string"
          }
        },
        "required": [
          "id",
          "segmentId",
          "proposedById",
          "anonymous",
          "payload",
          "state",
          "createdAt",
          "updatedAt",
          "objections"
        ],
        "type": "object"
      },
...
      "ReorderSegmentsDto": {
        "properties": {
          "orderedIds": {
            "description": "Segment IDs in their new order. Must list every segment of the event exactly once.",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "orderedIds"
        ],
        "type": "object"
      },
...
      "SegmentDetailResponseDto": {
        "properties": {
          "createdAt": {
            "example": "2026-05-26T08:12:00.000Z",
            "type": "string"
          },
          "description": {
            "example": "Outdoor garden venue within 30 minutes of the church, capacity 120+.",
            "nullable": true,
            "type": "string"
          },
          "eventId": {
            "example": "6bb69700-72eb-40a0-b37e-e6c0b8b004a9",
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "example": "e3b0c442-98fc-4e1b-9c2b-9d3a7c2f1a4d",
            "format": "uuid",
            "type": "string"
          },
          "lockedAt": {
            "example": "2026-05-28T14:33:12.000Z",
            "nullable": true,
            "type": "string"
          },
          "lockedById": {
            "example": "usr_2N9k8x4QyT3pA7c1bWv0eR",
            "nullable": true,
            "type": "string"
          },
          "lockedValue": {
            "additionalProperties": true,
            "description": "Final decided payload once the segment is locked. Shape varies by segmentType (e.g. a venue payload has { name, address, capacity }).",
            "example": {
              "address": "12 Marina Way, Lagos",
              "capacity": 150,
              "name": "Garden Court Hotel"
            },
            "nullable": true,
            "type": "object"
          },
          "position": {
            "description": "Zero-based ordering within the event.",
            "example": 2,
            "type": "number"
          },
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/EmbeddedProposalDto"
            },
            "type": "array"
          },
          "segmentType": {
            "description": "Free-form short slug ('date_time' | 'venue' | 'budget' | 'agenda' | 'guest_list' | 'roles' | 'theme' | 'travel' | …). Drives the picker icon on the planning board.",
            "example": "venue",
            "type": "string"
          },
          "state": {
            "enum": [
              "seeded",
              "open",
              "proposed",
              "objected",
              "converging",
              "single_remaining",
              "locked"
            ],
            "example": "open",
            "type": "string"
          },
          "title": {
            "example": "Reception venue",
            "type": "string"
          },
          "updatedAt": {
            "example": "2026-05-26T08:12:00.000Z",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "segmentType",
          "title",
          "position",
          "state",
          "createdAt",
          "updatedAt",
          "proposals"
        ],
        "type": "object"
      },
      "SegmentResponseDto": {
        "properties": {
          "createdAt": {
            "example": "2026-05-26T08:12:00.000Z",
            "type": "string"
          },
          "description": {
            "example": "Outdoor garden venue within 30 minutes of the church, capacity 120+.",
            "nullable": true,
            "type": "string"
          },
          "eventId": {
            "example": "6bb69700-72eb-40a0-b37e-e6c0b8b004a9",
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "example": "e3b0c442-98fc-4e1b-9c2b-9d3a7c2f1a4d",
            "format": "uuid",
            "type": "string"
          },
          "lockedAt": {
            "example": "2026-05-28T14:33:12.000Z",
            "nullable": true,
            "type": "string"
          },
          "lockedById": {
            "example": "usr_2N9k8x4QyT3pA7c1bWv0eR",
            "nullable": true,
            "type": "string"
          },
          "lockedValue": {
            "additionalProperties": true,
            "description": "Final decided payload once the segment is locked. Shape varies by segmentType (e.g. a venue payload has { name, address, capacity }).",
            "example": {
              "address": "12 Marina Way, Lagos",
              "capacity": 150,
              "name": "Garden Court Hotel"
            },
            "nullable": true,
            "type": "object"
          },
          "position": {
            "description": "Zero-based ordering within the event.",
            "example": 2,
            "type": "number"
          },
          "segmentType": {
            "description": "Free-form short slug ('date_time' | 'venue' | 'budget' | 'agenda' | 'guest_list' | 'roles' | 'theme' | 'travel' | …). Drives the picker icon on the planning board.",
            "example": "venue",
            "type": "string"
          },
          "state": {
            "enum": [
              "seeded",
              "open",
              "proposed",
              "objected",
              "converging",
              "single_remaining",
              "locked"
            ],
            "example": "open",
            "type": "string"
          },
          "title": {
            "example": "Reception venue",
            "type": "string"
          },
          "updatedAt": {
            "example": "2026-05-26T08:12:00.000Z",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "segmentType",
          "title",
          "position",
          "state",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "UnlockSegmentDto": {
        "properties": {
          "reason": {
            "description": "Optional human-readable reason for unlocking. Strongly encouraged for audit clarity.",
            "maxLength": 500,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateSegmentDto": {
        "properties": {
          "description": {
            "description": "Pass empty string to clear.",
            "maxLength": 1000,
            "type": "string"
          },
          "segmentType": {
            "maxLength": 32,
            "minLength": 1,
            "type": "string"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "cookie": {
        "in": "cookie",
        "name": "better-auth.session_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Event Planning Decision Making Assistant. Spec frozen per Phase 5.5 — bump `info.version` (major for breaking, minor for additive) on every committed change.",
    "title": "EPDEMA API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/events/{eventId}/segments": {
      "get": {
        "operationId": "segments_list",
        "parameters": [
          {
            "description": "Parent event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SegmentResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List segments for an event (ordered by position)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List segments for an event",
        "tags": [
          "segments"
        ]
      },
      "post": {
        "operationId": "segments_create",
        "parameters": [
          {
            "description": "Parent event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSegmentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentResponseDto"
                }
              }
            },
            "description": "Add a custom segment (admin only). Respects tier maxSegments cap."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a custom segment (admin)",
        "tags": [
          "segments"
        ]
      }
    },
    "/events/{eventId}/segments/{segmentId}": {
      "delete": {
        "operationId": "segments_delete",
        "parameters": [
          {
            "description": "Parent event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Segment UUID.",
            "in": "path",
            "name": "segmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a segment (admin only). Refused if the segment is locked or has live proposals. Returns 204 No Content on success."
          },
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a segment (admin)",
        "tags": [
          "segments"
        ]
      },
      "get": {
        "operationId": "segments_detail",
        "parameters": [
          {
            "description": "Parent event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Segment UUID.",
            "in": "path",
            "name": "segmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentDetailResponseDto"
                }
              }
            },
            "description": "Segment detail with live proposals + objections embedded. One call for the segment-detail page."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get segment detail with proposals",
        "tags": [
          "segments"
        ]
      },
      "patch": {
        "operationId": "segments_update",
        "parameters": [
          {
            "description": "Parent event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Segment UUID.",
            "in": "path",
            "name": "segmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSegmentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentResponseDto"
                }
              }
            },
            "description": "Rename or edit a segment (admin only). Works on locked segments too."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename or edit a segment (admin)",
        "tags": [
          "segments"
        ]
      }
    },
    "/events/{eventId}/segments/{segmentId}/lock": {
      "post": {
        "operationId": "segments_lock",
        "parameters": [
          {
            "description": "Parent event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Segment UUID.",
            "in": "path",
            "name": "segmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockSegmentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentDetailResponseDto"
                }
              }
            },
            "description": "Lock the segment with the chosen proposal as the winner. Atomic state transition under serializable txn + per-segment advisory lock."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lock a segment with a winning proposal",
        "tags": [
          "segments"
        ]
      }
    },
...
    "/events/{eventId}/segments/{segmentId}/unlock": {
      "post": {
        "operationId": "segments_unlock",
        "parameters": [
          {
            "description": "Parent event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Segment UUID.",
            "in": "path",
            "name": "segmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlockSegmentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentDetailResponseDto"
                }
              }
            },
            "description": "Unlock a locked segment. Restores all proposals to live state; clears lockedValue."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unlock a locked segment",
        "tags": [
          "segments"
        ]
      }
    },
    "/events/{eventId}/segments/reorder": {
      "post": {
        "operationId": "segments_reorder",
        "parameters": [
          {
            "description": "Parent event UUID.",
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderSegmentsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SegmentResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Reorder all segments in one atomic transaction (admin only). Must list every segment ID exactly once."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reorder all segments (admin)",
        "tags": [
          "segments"
        ]
      }
    },
...
    "/proposals/{proposalId}/withdraw": {
      "post": {
        "operationId": "proposals_withdraw",
        "parameters": [
          {
            "description": "Proposal UUID.",
            "in": "path",
            "name": "proposalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponseDto"
                }
              }
            },
            "description": "Withdraw a proposal (proposer only)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Withdraw a proposal (proposer)",
        "tags": [
          "proposals"
        ]
      }
    },
...
    "/segments/{segmentId}/proposals": {
      "get": {
        "operationId": "proposals_list",
        "parameters": [
          {
            "description": "Parent segment UUID.",
            "in": "path",
            "name": "segmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List live proposals for a segment."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List live proposals on a segment",
        "tags": [
          "proposals"
        ]
      },
      "post": {
        "operationId": "proposals_create",
        "parameters": [
          {
            "description": "Parent segment UUID.",
            "in": "path",
            "name": "segmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponseDto"
                }
              }
            },
            "description": "Create a proposal on a segment."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a proposal on a segment",
        "tags": [
          "proposals"
        ]
      }
    },
...
  },
  "servers": [],
...
}